FIRE-710 Refactor decorators directory and file name
