+++ src/friends/friends.controller.ts
@@ -14,9 +14,9 @@ import { ApiTags } from '@nestjs/swagger';
 import { User } from '@prisma/client';
 import { GetUserDecorator } from 'src/auth/decorator';
 import { JwtGuard } from 'src/auth/guard';
-import { ApiAuthDocument } from 'src/common/decorators/swagger/auth.document';
-import { API_DOC_TYPE } from './constants/docs.enum';
-import { FriendDocumentHelper } from './decorators/friend.document.decorators';
+import { ApiAuthDocument } from 'src/common/decorator/swagger/auth.document.decorator';
+import { API_DOC_TYPE } from './constants/friends-docs.enum';
+import { FriendDocumentHelper } from './decorator/friend-document.decorator';
 import { CreateFriendDto, DeleteFriendDto, UpdateFriendDto } from './dto';
 import { FriendsService } from './friends.service';
 import { FriendshipQueryParams } from './types/friendship.type';
