+++ src/records/records.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Delete, Get, UseGuards } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { GetUserDecorator } from 'src/auth/decorator';
 import { JwtGuard } from 'src/auth/guard';
-import { ApiAuthDocument } from 'src/common/decorators/swagger/auth.document';
+import { ApiAuthDocument } from 'src/common/decorator/swagger/auth.document.decorator';
 import { RecordsService } from './records.service';
 
 @ApiTags('records')
