+++ src/rooms/constants/rooms-docs.enum.ts
