+++ src/rooms/rooms.controller.ts
@@ -14,9 +14,9 @@ import { ApiTags } from '@nestjs/swagger';
 import { User } from '@prisma/client';
 import { GetUserDecorator } from 'src/auth/decorator';
 import { JwtGuard } from 'src/auth/guard';
-import { ApiAuthDocument } from 'src/common/decorators/swagger/auth.document';
-import { API_DOC_TYPE } from './constants/docs.enum';
-import { RoomsDocumentHelper } from './decorators/rooms.document.decorators';
+import { ApiAuthDocument } from 'src/common/decorator/swagger/auth.document.decorator';
+import { API_DOC_TYPE } from './constants/rooms-docs.enum';
+import { RoomsDocumentHelper } from './decorator/rooms-document.decorator';
 import { CreateRoomDTO, GetRoomDTO } from './dto';
 import { RoomsService } from './rooms.service';
