+++ src/users/decorator/user-document.decorator.ts
@@ -1,0 +1,3 @@
+import { applyDecorators } from '@nestjs/common';
+
+export function UsersDocumentHelper() {}
